Sistem/naval/apps/src/requirements.txt
absl-py==1.4.0
aioice==0.9.0
aiortc==1.5.0
ajsonrpc==1.2.0
altair==4.2.0
anyio==3.7.1
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
autopep8==2.0.0
av==10.0.0
backcall==0.2.0
bidict==0.22.1
blinker==1.6.2
bottle==0.12.25
branca==0.6.0
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.10.0
Cython==3.0.0
decorator==5.1.1
deprecation==2.1.0
dnspython==2.4.0
easyocr==1.7.1
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
et-xmlfile==1.1.0
execnet==1.9.0
executing==1.2.0
filelock==3.10.7
filterpy==1.4.5
firebase-admin==6.1.0
Flask==2.3.2
Flask-SocketIO==5.3.5
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.39.2
gast==0.4.0
gcloud==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==2.10
ifaddr==0.2.0
imageio==2.26.0
importlib-metadata==5.1.0
imutils==0.5.4
iniconfig==2.0.0
ipython==8.11.0
itsdangerous==2.1.2
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
jws==0.1.3
kadita==0.0.2
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.1
marshmallow==3.21.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe==0.10.3
ml-dtypes==0.2.0
mock==5.0.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.5
msgpack-numpy==0.4.8
mss==9.0.1
mutagen==1.46.0
mysql-connector-python==8.3.0
networkx==3.0
ninja==1.11.1.1
numpy==1.25.1
oauth2client==4.1.3
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
path==16.6.0
path.py==12.5.0
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
platformio==6.1.13
playsound==1.3.0
plotly==5.16.0
plotly-express==0.4.1
pluggy==1.0.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydeck==0.8.0
pyee==11.0.0
pyelftools==0.30
pygame==2.5.1
PyGetWindow==0.0.9
Pygments==2.13.0
PyJWT==2.7.0
pylibsrtp==0.8.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRect==0.2.0
pyrsistent==0.19.2
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
pyTelegramBotAPI==4.12.0
pytest==7.3.1
pytest-shutil==1.7.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-firebase==1.2
python-jwt==2.0.1
python-socketio==5.8.0
python-telegram-bot==20.4
pytweening==1.0.7
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==12.6.0
roboflow==1.1.6
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
semver==2.13.0
sentry-sdk==1.17.0
shapely==2.0.3
simplejpeg==1.6.6
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.30.0
statsmodels==0.14.0
streamlit==1.25.0
streamlit-autorefresh==1.0.1
streamlit-custom-notification-box==0.1.2
streamlit-js-eval==0.1.5
streamlit-modal==0.1.0
streamlit-option-menu==0.3.6
streamlit-webrtc==0.45.1
supervision==0.14.0
sympy==1.11.1
tabulate==0.9.0
tclab==1.0.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tflite==2.10.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.15
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ultralytics==8.1.23
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.23.1
validators==0.20.0
vidgear==0.3.0
watchdog==3.0.0
watchgod==0.8.2
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.14.1
wsproto==1.2.0
yt-dlp==2023.7.6
zipp==3.11.0
